move common cli bits to module instead of crate
