+++ Cargo.toml
@@ -2,5 +2,4 @@
 members = [
     "cargo-espflash",
     "espflash",
-    "espflash-common",
 ]
