--- espflash-common/README
@@ -1,1 +1,0 @@
-Common utilities for `espflash` and `cargo-espflash` cli commands
